Sql rework c# side, broken things revisioned
After the sql rework, i have also reworked most of the c# scripts connecting to it within unity, and now that's (sort of) done, i have started fixing the newly broken scripts that was the result of the rework. as of now, that is the tower buying script and the menus getting to it.